tests in p2p compile but are broken, need to work on handshake nil pointer
