+++ p2p/transport.go
@@ -33,7 +33,6 @@ import (
 const (
 	maxUint24 = ^uint32(0) >> 8
 
-	sskLen = 16                     // ecies.MaxSharedKeyLength(pubKey) / 2
 	sigLen = crypto.SignatureLength // elliptic S256
 	pubLen = 64                     // 512 bit pubkey in uncompressed representation without format byte
 	shaLen = 32                     // hash length (for nonce etc)
